refactor(Video): add a VideoPlayer folder
